Unify the assessment selection modal (with some renaming)
