+++ app/views/landing/show.html.erb
@@ -1,6 +1,6 @@
-<div data-controller="country-selection">
+<div data-controller="assessment-selection">
   <%= hidden_field_tag "countries_and_assessments", @selectables.to_json, data:
-  { target: "country-selection.selectables" } %>
+  { target: "assessment-selection.selectables" } %>
   <div class="landing container">
     <div class="row">
       <div class="col left-panel">
@@ -22,12 +22,16 @@
         <%= form_tag '/country_select', id: "country-select-menu" do %>
           <div class="input-group">
             <%= select_tag "country",
-            options_from_collection_for_select(@countries, 'name', 'name'), class:
-            "custom-select", data: { target: "country-selection.selectedCountry" }
+              options_from_collection_for_select(@countries, 'name', 'name'),
+              class: "custom-select",
+              data: { target: "assessment-selection.selectedCountry",
+                      action: "assessment-selection#selectCountry" }
             %>
             <div class="input-group-append">
-              <%= button_tag "Get Started", type: :button, data: { action:
-              "country-selection#select" }, class: "btn text-white bg-green" %>
+              <%= button_tag "Get Started",
+                type: :button,
+                data: { action: "assessment-selection#openModal" },
+                class: "btn text-white bg-green" %>
             </div>
           </div>
         <% end %>
@@ -36,5 +40,5 @@
       <div class="col"></div>
     </div>
   </div>
-  <%= render "shared/country_selection_modal" %>
+  <%= render "shared/assessment_selection_modal" %>
 </div>
