+++ app/views/shared/_assessment_selection_modal.html.erb
@@ -2,13 +2,17 @@
     class="modal"
     tabindex="-1"
     role="dialog"
-    id="joint-country-selection-modal"
+    id="assessment-selection-modal"
   >
   <div class="modal-dialog" role="document">
     <div class="modal-content">
       <div class="modal-header">
         <h1 class="modal-title display-header mx-4">
-          LET'S GET STARTED!
+          Let's get started on
+          <span
+            id="country-name-header"
+            data-target="assessment-selection.selectedCountryName"
+          />
         </h1>
         <button
           type="button"
@@ -36,23 +40,21 @@
           from technical areas.
         </p>
 
+        <p><b>Where would you like to start?</b></p>
+
         <%= form_tag '/country_select', id: "assessment-select-menu" do %>
-          <%= hidden_field_tag "country", nil, data: { target: "joint-country-select.selectedCountryModal" } %>
-          <div class="input-group">
-            <%= select_tag "country",
-              options_from_collection_for_select(@countries, 'name', 'name'),
-              class: "custom-select",
-              data: { target: "joint-country-select.selectedCountry",
-                      action: "joint-country-select#select_country" }
-            %>
+          <%= hidden_field_tag "country",
+            nil,
+            data: { target: "assessment-selection.selectedCountryModal" }
+          %>
           <div class="input-group mb-2">
             <%= select_tag "assessment_type",
               options_from_collection_for_select([], 'assessment_type', 'assessment_type'),
               class: "custom-select",
-              data: { target: "joint-country-select.assessmentTypes" } %>
+              data: { target: "assessment-selection.assessmentTypes" }
+            %>
           </div>
-          <%= button_tag "Next", type: :submit, class: "btn btn-primary w-100" %>
-        <% end %>
+        <%= button_tag "Next", type: :submit, class: "btn btn-primary w-100" %> <% end %>
       </div>
     </div>
   </div>
